+++ gcovr/coverage.py
@@ -137,6 +137,8 @@ class BranchCoverage:
             Whether the branch is excluded.
     """
 
+    first_undefined_blockno: bool = True
+
     __slots__ = (
         "blockno",
         "count",
@@ -169,7 +171,9 @@ class BranchCoverage:
     def blockno_or_0(self) -> int:
         """Get a valid block number (0) if there was no definition in GCOV file."""
         if self.blockno is None:
-            LOGGER.info("No block number defined, assuming 0")
+            if BranchCoverage.first_undefined_blockno:
+                BranchCoverage.first_undefined_blockno = False
+                LOGGER.info("No block number defined, assuming 0 for all undefined")
             return 0
         return self.blockno